archive_crawling (as on Neo)/requirements.txt
alchemy-mock==0.4.3
bs4==0.0.1
datefinder==0.7.1
dateparser==0.7.6
dotmap==1.3.17
html5lib==1.1
langdetect==1.0.8
latest-user-agents==0.0.3
lxml==4.5.0
mongomock==3.22.1
mysql-connector==2.2.9
newspaper3k==0.2.8
pandas==1.2.0
pymongo==3.11.2
python-dateutil==2.8.1
pytz==2020.4
readability-lxml==0.8.1
requests==2.24.0
requests-mock==1.8.0
retrying==1.3.3
sqlalchemy==1.3.20
urllib3==1.25.8
